+++ plugins/disrumpo/tests/unit/display_format_test.cpp
@@ -13,6 +13,7 @@
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/catch_approx.hpp>
+#include <cmath>
 
 #include "plugin_ids.h"